online_inference/requirements.txt
dataclasses==0.6
Faker==8.1.3
requests==2.24.0
pydantic==1.8.1
numpy==1.19.2
uvicorn==0.13.4
pytest==6.2.4
marshmallow_dataclass==8.4.1
omegaconf==2.0.6
pandas==1.1.3
fastapi==0.65.0
PyYAML==5.4.1
scikit-learn==0.23.2
